Correct check for errors in applicationService
Previously we were checking for the `length` method to be truthy on the
errors object. This would always be falsy as we don’t define a `length` on
the errors object.
I assume that this was an array but at some point but it changed to being
an object. TS didn’t highlight this error because the check for ‘.length’
on a `Record<string, string>` type is valid.