Add page object for RehabilitativeInterventions
There is a bit of nastiness here in the ‘items’ method.
We want to keep as much of the copy as possible in the rehabilitativeInterventions.ts
file so it can be unit tested in order for changes to break tests so
that we remember to update things. However the ‘other’ checkbox needs
to trigger a textbox to show which means it needs to use a variable
which can only be in nunjucks land.
We need the translation to be in TS though as that’s where our translations
live so I have opted to have it in the object on line 7 so we can use it
for the Interventions type and responses translations but we need to remove
it in the ‘items’ method. It’s not ideal but hopefully it’s clear enough
what’s going on.