Set up known state in e2e tests
After the attempt in #304 to fix the UI tests, it seems that the
extension tests are still a bit flaky. The combat this, I thought it
might be best to set up a known state in the application, rather than
relying on a state that may or may not exist. Now we create a booking
and then attempt to extend that exact booking. It will make the tests a
tad slower, but hopefully more reliable!