Invoke `page.previous` function in evaluation in layout.njk
Previously we were checking for the existence of the function.
This would always evaluate to true.
This means where we previously used ‘’ as a stand in for navigating back to
the tasklist the back button will no longer show - however this functionality
was broken. I think it would be useful to be fixed in future but outside
of the scope of this PR.