Correct assignments from user response for placementPurpose and rehabilitativeInterventions
It is possible that a user could not respond to either of these questions.
In this case we would add 'undefined' to the array. We had handled this error
however this change makes things more robust and less prone to error in future.
EG a future maintainer may assume an array with length 1 would contain
a response however that 1 item could be ‘undefined’.