Add a generator to generate form pages
This is very scrappy and procedural, but will do the job for us. It
generates a skeleton page class, a test file, a view, and a page object
for use in Cypress and adds `TODO` comments for all the things that need
adding.

The only thing it doesn’t do is add a new section to the imports, but
this requires manual work anyway. It does add a TODO as a reminder to
do it manually.