Add an `applyInput` macro
This wraps around the `govukInput` macro and ensures we don’t have to
repeat ourselves when adding the id and name, as well as ensure the
error message is added, as well as  pre-populating the field with text.

We have the pass the context into the macro too, as fetching the context
within the macro itself only fetches the context of the macro, not the
page we call it from.