Update routing
If the user specifies any additional needs on the accessNeeds page
then they are taken to AccessNeedsAdditionalAdjustments.
The exception to this is the user answers ‘none’ and only ‘none’.
They could, rather paradoxically, answer ‘none’ and give an additional
need. In this case we assume it’s a mistake and navigate them to either
the mobility needs page or accessNeedsAdditionalAdjustments page