Page class for AccessNeedsAdditionalAdjustments
The getNeeds method isn’t as neat as I would like but it works.
It adds commas after each listen item except if it is the only item
in the last OR if it is the penultimate or ultimate item in the list.

Another thing to note here is the tests for the previous and next page.
For some reason (probably due to us wrapping our own describes to reduce
duplication) we have to explicitly recreate the application object in its
own describe block before passing it to these tests.