Add state to Apply page objects
This adds a JSON fixure with completed application data. We then use
this fixuture to create an application which is then passed to each
page object and use the data to initialize a Apply Task Page class with
the data taken from the fixture.

This will allow us to then keep a record of what fields have been
filled in, and then use the translations in the Task Page to confirm
the check your answers page is populated correctly in a future commit.