Add the Check Your Answers page
This adds a view for the check your answers page, as well as a link to
the page from the task list. The integration tests gather the page
objects we’ve built for the various steps in arrays. When we hit the
check your answers page, we then loop through each page object, fetch
the Form Page class that’s associated with the page object, and assert
that each item in the Form Page response is included in the correct part
of the page for each section.