Move response-related functions to helpers
This makes the ApplicationService leaner, and also allows us to use
the helpers in views without relying on the service. I’ve also tweaked
the `getResponsesAsSummaryListItems` helper to return an array of
sections with their title and their repsonses as summary list rows.
This will make it easier to use within a view.