Update the async setup method
This is the first time this approach has been used in anger, and it’s
transpired that having an async `setup` method that gets used after the
class gets initialized is not enough, as there will be data that gets
fetched that we’ll need during initialization.

With this in mind, I’ve changed this approach to add a static async
`initialize` method that we check for the existence of. If it exists,
then we call the relevant data services, marshal the data into the
correct format, pass it onto the initializer and then return the class
instance.

We also now don’t pass the whole request object, just the request
body, the token, the application object, and the data services. This
makes testing easier and solves some issues we were having with
using the page classes in the Cypress tests.

This has also been an opportunity to refactor the tests a bit. We can
now stub the `getPage` function, as it have moved to the
applicationUtils. This means I’ve had to move some of the tests over to
the applicationUtils tests, as the functionality we were originally
testing was in the applicationServices tests.

That’s a lot of words for a few lines of code! If you’re still reading
this - congratulations!