Display array of objects as summaryList
This updates the checkYourAnswersUtils to check the type of a response.
If it’s a string, we just return a the response as it is, if it’s an
array of objects, we cycle through the array and the key/value pair
to generate a summary list as HTML. I’ve added some styling to make an
embedded summary list appear slightly more compact. This can be built
on as we iterate. I’ve also checked it over with @beth-dixxon

I’ve also added some Cypress tests to check the response appears as we
expect in the check your answers page.