Add Adjudications to the response
And also test they appear on the Check Your Answers page.

This uses a similar technique to the checkboxes work in an earlier
commit, with two helper functions that we test seperately. When
we call the `details` method we can just check the helper function
has been called, rather than having to set up a bunch of fixture
data. This makes the tests tidier and easier to understand.