Remove table-related methods from case notes
This removes the table-related functionality from the caseNotes class,
and moves the responsiblity for rendering the table directly to the
view.

The only exception is the `checkBoxForCaseNoteId` method, which fetches
a case note by ID, checks if it is in the `selectedCaseNotes` array, and
then calls a helper function in the same file. This means we can test
the HTML seperately from the class, and when we test the class directly,
all we need to do is check the class calls the helper method, rather
than painstakingly checking each bit of HTML and setting up / tearing
down a lot of test data.

I’ve also added a `toMatchStringIgnoringWhitespace` Jest matcher, making
it easier to compare strings of HTML without making sure the whitespace
is exactly identical.