Return user to tasklist after answering oral hearing question
Previously we took the user to the Placement Date page which requires
there to be a release date in the session however on the Release Date
page they answered that they don’t know the the release date so it
doesn’t make sense to move them on to the placement date page. So instead
we return them to the tasklist.
This has been confirmed by our SME here: https://mojdt.slack.com/archives/C02SHKETZ7F/p1668418539408919?thread_ts=1668174667.005489&cid=C02SHKETZ7F