Add getArrivalDate function
We need to be able to retrieve the arrival date from an application.
This is potentially risky if the schema changes so I’ve added a lot of
error handling to help us point us in the direction if something goes
wrong.
I would have loved to replace the ‘fetchArrivalDate’ method in
the server/form-pages/apply/move-on/placementDuration.ts file with this
function but unfortunately it would introduce a cyclical dependency.
Something to do in future through