Add a script to generate the schema
This goes through each of the pages, and uses `typescript-json-schema`
to convert the body type to a JSON schema, it then marshals it into the
correct format and writes it to a file.

Each form has a `generateSchema` file which calls the schema generator
with the pages and a type. This is only for Apply at the moment, but
we’ll leverage this for Assess soon too