Fetch risks alongside assessments
As we currently have to fetch risks via a seperate endpoint, we need to
add a new `AssessmentWithRisks` type, and then loop through each
assessment, fetching the risks as we go. This will hopefully go away
with some API work that is scheduled, but this gets us part of the way.