Send the index offence to the API
This gets the index offences from the API again, and then selects the ID
of the selected offence. We could have passed through each individual
element (convictionId, deliusEventNumber, offenceId) as individual
fields, but I feel this gives us extra confidence and flexibility when
we have to select from a number of index offences in the future.