Redirect to offences if an offence is not provided
If an offence has not been provided in the previous step (i.e. there is
more than one), then we redirect to the selectOffence screen. I’ve also
tweaked the integration test to only return one offence in the happy
path. I’ll add a new integration test to test the new path in a future
commit.