Add an attachDocuments page class
This has an initialize function, which fetches the documents from the
relevant endpoint, filters the selected documents from the body
(if provided) and adds the total list of documents to the class
instance.

The response is a simple key/value record with the filename as the key,
and the description as the value (with `No Description` as the default).
This reflects what we have in the Check Your Answers prototype designs.