Ensure application we’re fetching has the same ID
When we’re loading the tasklist, and the user already has an in-progress
application and then attempts to load another application, the app tries
to get the application from the session, which will always be the
in-progress one. We already have this functionality in the
applicationService, so we’ll make it a public method, and use it in the
show action (which loads the tasklist)

It’s an open question as to whether this session stuff is useful, as
we’re sending data to the API each time now, but this fixes the bug
for now.