Add rate limiting to the authentication endpoints
Following a recommendation by CodeQL, I’ve added some rate limiting to
the endpoints that use authentication. This uses `express-rate-limit`
and `rate-limit-redis` to prevent too many requests to specific
endpoints/middleware. The rateLimiter function has options we can pass
to set the window and max number of requests that can be made to an
endpoint before hitting a rate limit.

For the authentication endpoints, I’ve set this to 10 requests per IP
address per 10 minute window, and 100 requests per 1 minute windw for
the authorisation endpoints which seems like a reasonable default to
start with.