Handle when Oasys Sections response has objects
When we revisit the Oasys Sections page (for example, to change a
response), the body contains the actual section objects, rather than
the section IDs, before, we would get an error as the mapping was
attempting to match the selected sections against the section ID of
the returned sections. This refactors the code to only do the mapping
if the section IDs are present, otherwise pass the objects through as
they are.