Add page class for reviewing application
This is the first page class for an assessment so there is some additional
boilerplate.
We could perhaps do without ‘tasks’ for Assessments but I opted to keep
them in for now.
I am not sure if we need the result of the ‘response’ method on the page
class so I haven’t added it for now but it will be easy to do if
required later.