Add a helper to complete the Apply journey
This will allow us to share the code used to complete the journey across
both the e2e and integration tests. There is also a
`setupApplicationStubs` method which sets up the stubs and the
expectations that will be used in the integration tests. This will not
be used in the e2e tests, so the expectations will be empty.