Add e2e tests for Apply
This uses the helper in the e2e tests.

The confirmation screen doesn’t currently work, as we have no staff
in the database to allocate an application to, so is commented out with
a todo.

I’ve also added a person fixture, which contains all the information we
expect the user with the CRN `D002399` to have.