Add download links for files attached to application review
The reviewer needs to be able to download the attached files.
This means that the attach documents section of the review needs to be
slightly different to all the other sections.
This isn’t the most elegant solution but it works. When we are generating
the rows for the review we generate all of them the same way *unless*
the page name is ‘attach-documents’ in which case we add the hyperlinks
for the user to download the documents.
In this commit we also add E2E tests to ensure the download works in
browser