Add Initialize method to Review Page Class
We need to enrich the application with the full document details so we
can add a link to the document on the review page.
In order to do this we fetch all the documents for the application.
We then iterate through the selected documents attached to the application
and collect the full document objects for the selected documents from the
documents we fetched from the API as we need the ‘id’ of each document.
We then replace the original ‘selectedDocuments’ array in the application
with our newly generated ‘selectedDocuments’ array.