Use an AssessHelper to complete the form
This adds a helper to be used in the Integration tests (and later the
e2e tests). To help us better simulate a real journey, I’ve added
an `updateAssessmentAndStub` method, which fetches the data from the
last completed step and uses it to populate the assessment data and
update the stub. This way we can check the status of the task as each
step is completed, without having to load an assessment from the
session, as we do in Apply.

If this works well for us in Assess, we can backfill this approach for
Apply too.