Allow full error trace in Nunjucks
When a global function in a template throws an error, we don’t see the
full trace of the error, which makes diagnosing issues really hard.

This makes use of the [undocumented dev argument](https://github.com/mozilla/nunjucks/issues/1430)
which allows us to see the cause of the error in our logs and in Sentry.

I’ve had to extend the nunjucks types to allow for this option to be
set, as it’s not in the type definition in `DefinitelyTyped`. However,
I’ve [opened a PR to add this](https://github.com/DefinitelyTyped/DefinitelyTyped/pull/63968)