Add PageClass for Matching Information
This page collects a wealth of information so it is incredibly meaty.
Essentially the information we are needing to get from the user is
their preferences for the accommodation for their PiP which comes in as
the result of many radio inputs (and a select).
There may be smarter ways of doing this but I (hopefully) have optimised
for readability.