Create radio matrix table function
This page needs a couple of tables that allow the user to choose a different
radio button for each row. There isn’t anything close to this in the GDS
or MoJ design systems so we’re going to have to make one. We also add
this to the nunjucks globals so we can use it in the view