Tidy up integration tests
This pushes a lot of the setup to the `beforeEach`, and sets up the
`assessHelper` and `application` as a shared variable.

It also tidies up some of the `then` operators, so we’re not nesting
multiple `then`s within each other.

There’s definitely room for improvement, but this makes the tests a bit
more readable