Add controller method for submitting assessments
If the checkbox on the taskslist isn’t selected we have to do some
manual validation. As the call to this method lives outside of the
simple ‘form-pages’ framework *and* the validation  is done in the UI
we have to create some error arrays and objects in
the format expected by the nunjucks macros.

If the checkbox is selected then we submit the application