Allow getResponses to be used for assessments too
We need to be able to use this function for Assessments as well so the
types needed to be updated. Thankfully the rest of the function can
be used wholesale.
It is perhaps a little misleading to have it in the ‘applicationUtils’
folder but there doesn’t seem to be a better place for it for now.