Add a getTaskStatus helper function
This accepts an application or assessment and a Task class. We find the
data for that task in that application / assessment. We first find the
first page with a response within that task (the first page in the task
may not be the first page in the journey).

If there’s no  data, we know the task hasn’t been started, so we return
a status of `not_started`.

We then loop though each page within the task. We then initialize the
page and check for errors. If there are errors, then we know the
task is in progress, but not complete, so we return an `in_progress`
status. If there are no errors, we get the next page. If there is no
next page, then this page is the last page in the task, and the
task is complete. If there is a next page, we rinse and repeat!