Add a TasklistService
This adds a class that accepts either an application or an assessment,
and then goes through each task within the relevant sections to get the
status for that task. If the previous task is anything other than
complete, we immediately record that task’s status as `cannot_start`.

With these statuses recorded in the initializer, we can then reliably
get the count of completed sections, as well as returning information
about each section, their tasks, and each task’s status.