Add a helper for assessment table rows
The code we have for matching assessment table rows is pretty
repetitive, so I’ve updated it to add a helper that runs the
table row function that we use in the view, turns it into an
array. We then loop through that array and check each table row
matches the table row data for each assessment.