Fall back to display document type description when extendedDescription blank
The `description` field is taken from the `Document.extendedDescription` which often
contains an excellent description such as "Non Statutory Intervention for
OPD Community Pathway on 23/08/2016". However it can be blank. In that
case it's important to show the applicant more than the document name which
in many cases is not very helpful.

A document name might be `e davey.pdf`. If there is no `description` to supplement
this meagre information we now supply the `typeDescription`, perhaps "PNC previous
convictions" which indicates that this is an important document to include
in the application.