Explaining where prison information is pulled
We've included content to show where the prison information is being pulled from. 'This information is imported from NOMIS'.

We've decided not to include the date the information is pulled in at the moment as it will always show the current date. We should keep track of this to see if users want to see the date in the future.