Catch a 403 error from the createApplication method
The API will return a 403 if the CRN is not in the users caseload.
This commit shows an error when the 403 is thrown and informs the user
that the CRN is not in their case load.

The redirect was not occurring properly without us returning the
‘catchAPIErrorOrPropogate’ function which then meant (due to linting rules)
we have to return from every path of the function. This then made the
‘else’ clause redundant which I think makes the function a little clearer