Contingency plan partners page class
This is a page in the format of the other Apply/Assess pages.
However there are a few interesting things about the class.
The designs necessitate that the user can add a partner agency, press
a button and then remain on the same page in order to add further partner
agencies. This isn’t something we’ve had to deal with before.
In order to do it we need to keep track of the `partnerAgencyDetails`
and pass this in to the view and have it returned to us.
We also need to be able to see when a partner agency has already been added to
the array as when the user submits an agency they are redirected back to the
page and that request is handled by the ‘show’ method of the pagesController
with the same body.