Contigency plan partners view
We have two use two seperate forms here in order to determine if the
user wants to add a new partner or to save and continue.
When the user saves and continues the hidden input ‘saveAndContinue’ is
also submitted along with the form which the page class uses to allow
the user to continue in the form.

Testing this page doesnt fit neatly into our testing patterns. As we
submit the page 3x (2x Partner Agencies, 1x ‘Save and Continue’) we have
to add it to the `numberOfPages` function 3 times. I feel like this
deserves a comment explaining why. Further as the ‘response’ shown in the
CYA is an array of objects we need to test it in a similar way to the
way we do caseNotes, adjudication and acct alerts.