Util func for deciding if the contingency plan pages should show
We may need to move this in future and as a standalone function it is
easier to mock. It needs to live in its own file to avoid dependency
cycles.

It is a fairly straightforward function complicated slightly by the
fact that applications only have a ‘release type’ response given certain
responses to the ‘sentence type’ page

I also add a function for adding a response to an application as this
code was being duplicated in a few places