Use ReleaseTypeOption type for ReleaseTypes
This ensures we have one source of truth for the possible ReleaseTypes.
The newly created `in_community` option is not used in the UI itseld, so
we create a new `SelectableReleaseTypes` type in the page class to
encompass all release types other than `in_community`. I’ve also updated
some of the naming in the page class to (hopefully) make things a bit
clearer