Ensure false items don’t get stripped out
We have a `filterBlanksFromData` method which strips out any empty
variables that are passed to the API. However, this also strips out any
literal `false` values, which we do want to send.

This updates the `filterBlanksFromData` helper to check if the item is a
literal boolean before checking it is ‘falsy’ (i.e. null, undefined or
a blank string). This ensures any literal `false` values are sent to the
API as expected.

I’ve also added a check to ensure all the expected keys are passed to
the API when submitting an application (where the error was first
identified).