Add difference in days util function
This will help us display the difference between dates in days consistently
by creating an object with two properties we will need when creating
tables which show date differences.
The first is a UI date that is for display to the user EG ‘3 days’.
The second is the difference as an integer which can be used for sorting.