Update the Integration tests
The expands the Assessment tests to check the right data is sent to the
API. I’ve changed the tests to mark the assessment as accepted, so we
can ensure the data is sent correctly. In an ideal world, we should add
a test to check the rejection information is sent, but I think we need
to make some tweaks to the journey before we do that.