Set right next page in SufficientInformation page
This was set incorrectly, so when we were walking the sections and tasks
the SufficientInformation page was showing as incomplete when
suffiecient information had not been recieved. I’m not _entirely_ sure
why the integration tests were passing, but it seems to work in both
cases now, so I’ll leave getting in the weeds too much for now.