Run tests in Parallel
This breaks all the parts of our test suite into constituent parts, so
we can run them in parallel. It doesn’t save a ton of time (maybe 2
minutes at the most), but ensures we get quick feedback from any
failures, and keeps things just that little bit tidier.

We also share a bunch of the setup in a new shared action, which
installs the dependencies, generates the types, and runs typecheck.