Both integration tests require wiremock
For me running `npm run test:integration` blew up with various errors including:

```
From Node.js Internals:
      at <unknown> (      Error: connect ECONNREFUSED ::1:9999)
      at TCPConnectWrap.afterConnect [as oncomplete] (node:net:1494:16)
```

This (via Docker Compose) points to the absence of a wiremock instance.

If we take a look at `npm run test:integration:ui` this does start wiremock. This makes sense as we expect our integration tests to be mocked out. If after running this you go back to `npm run test:integration` it will magically start working again! This is due to the fact the wiremock container wasn't removed and so allows the tests to work accidentally.