Correct response from allocations endpoint
I had assumed the response from the  /applications/{applicationId}/allocations endpioint would be an assessment object which would contain the name of the person it has been assigned to. However it returns a User object.
We could look up the user by the ID returned but I want to fix ASAP and can improve later.