Update the task show action
This updates the show action to support multiple task types. At the
moment, we’re still fetching only assessors with qualifications for a
particular application. I feel like this should now be a responsibility
of the backend, and showing a task should return a list of qualified
users too, but in the interests of getting this working, I’ll leave it
like this for now and make a card to do the backend changes.