Create utils to map between UI and API BedSearchParams
The version from the query string will always be a string whereas
the API can only accept a number so we have to map.
I would have preferred to have ’mapApiParamsForUi’ as a method on the
BedSearchParams factory but it doesn’t appear to be possible for a
Fishery Factory to return two different types.