Attempt at fixing bookingService test flakiness
This test has intermittently failed since it was written but there
is nothing obviously wrong with it. The failures seem to have been
getting more frequent recently (presumably because of the large number
of tests we have).
My guess is that as the booking object is relatively large it is
taking a long time to be built but somewhere (outside of our code)
it is not being awaited properly. I could well be wrong though!
So we set up the booking objects before the test runs and reduce the
number of bookings it creates. Locally this seems to have fixed the issue.

Attempt at fixing bookingService test flakiness

This test has intermittently failed since it was written but there
is nothing obviously wrong with it. The failures seem to have been
getting more frequent recently (presumably because of the large number
of tests we have).
My guess is that as the booking object is relatively large it is
taking a long time to be built but somewhere (outside of our code)
it is not being awaited properly. I could well be wrong though!
So we set up the booking objects before the test runs and reduce the
number of bookings it creates. Locally this seems to have fixed the issue.