Add check for query when rendering Search template
This is necessary as after the initial render the template is then
rendered again and express doesn’t register the query params - for
reasons I am not entirely sure of.
This conditional prevents an error due to the lack of query params
and makes sure the page renders properly. I have borrowed the approach
from the Temporary Accommodation repo.