Add search inputs
This adds inputs to allow the user to search for beds by start date,
distance, length of stay and location.
This adds some complexity to the controller as we have to take both the
query string params and the body params and use both of them (if they both exist)
with the body params taking precedence over the query params.