Refactor `assertDefinition`
The old way of doing things was veeeery slow, and meant that we were
searching every single `dt` and `dd` by their text, which I think was
the cause of some CI performance issues. This gets the `dt` and then
checks the text content of its sibling `dd` element which _should_ be
quicker.