Map a placement request into bed search params
We need to convert the data we have in the Placement Request into a format
that the API Bed Search endpoint accepts. Rather than do that in the
Bed Search Controller I think it makes sense to do it at this point.
In the UI we don’t have an idea of whether the ‘essentialCriteria’ relates
to the room or the premises so for now we map it to both ‘requiredRoomCharacteristics’
and ‘requiredPremisesCharacteristics’. Eventually this will be fixed
by an API side change that will allows us to simply search for ‘requiredCharacteristics’.
Some properties of a placement request (EG ‘mentalHealthSupport’) will
also be covered in these characteristics