`oasysImportReponse` returns question names
Before, there was an `Array.isArray` check, which never got called in
reality, as the question responses will always be objects. As a result,
all we were returning in the response was a Record with the question
number as the key and the response as the value. This updates the method
to fetch the summary by its question number, and return the
human-readable question name to use that as the key. This makes for a
prettier and more readable summary in the Check Your Answers pages.

I’ve also updated the typing of all the page classes, as well as the
schema. The application data (which is used in the integration tests)
has also been updated to reflect what the application data really looks
like.