Show navigation in the header
We also need to add a small bit of middleware to allow us to get the
path the user is currently on. This also means we have to add a user
local to the errorHandler tests, as these test the full stack, so we
need a user to query roles on.