Add to BedSearchParametersUi type
We have a BedSearchParameters type generated from the API spec which
is the parameters we should be passing to the API.
BedSearchParametersUi however deals with the parameters that are passed
to the bed search controller ‘index’ method.
Previously these properties had only been added to the BedSearchParams
type in the mapPlacementRequestToBedSearchParams function which meant
we couldn’t use them in the BedSearchController but by adding them
to the BedSearchParametersUi type we are able to use them more widely.