Add view and integration tests for the show page
I’ve also added a `shouldContainSummaryListItems` helper to check a
page contains an array of summaryList items. A similar thing was used in
the read-only applications test, so I’m using that there as well. The
only change is I’m using the native JS browser `DOMParser` rather than
jQuery as some of the HTML content was choking on special characters.