Export 'ApType' type
We want to make sure that the types we use in the placementDuration
page match up with those we collect in the ApType page so we will
use the type from ApType.
There is another ApType type in the shared types generated from the API
spec but unfortunately this uses different terms from what is collected
in the UI (‘normal’ in the API instead of ‘standard’ in the UI).
We also rename the page class imports to be SelectApType in the index
to match what it is called in the apType.ts file to avoid a collision
with the type name