Define UI parameters seperately
I found it a bit confusing, so I felt it was easier to be more
prescriptive, and have a separate factory too. As part of this, I’ve
removed the `[index: string]: unknown` from the `BedSearchParametersUi`
interface, as this made the type behave like a record, so could have any
additional keys. To get around this in the controller I’ve updated
`objectIfNotEmpty` to accept the object that’s set in the generic.