Send `durationWeeks` in the UI
The UI already said the user was entering weeks, and the duration of the
placement is stored in the placement request object as weeks, so this
was incorrect. As part of this I’ve had to update the params converters
in matchUtils to convert from days to weeks and vice versa.