Add method to get Update Data for an application
The new `UpdateApprovedPremisesApplication` type is has the same
first-class fields as the `SubmitApplication` type, but all the
attributes are non-mandatory. This combines the inital
`applicationSubmissionData` method and helpers into a new file, which
also includes a `getApplicationUpdateData` method, which uses the same
helpers, but calls `retrieveOptionalQuestionResponseFromApplicationOrAssessment`
instead of `retrieveQuestionResponseFromApplicationOrAssessment`,
meaning missing values will not throw an error.