Change aria-allowed-attr to not fail test
The govuk-frontend package we use is generally considered a best practice UI toolkit.
However it includes the ‘aria-expanded’ attribute on radio buttons which
currently goes against WCAG guidelines. GDS, the maintainers of govuk-frontend,
have petitioned WCAG to allow this attribute on radios: https://github.com/w3c/aria/issues/1404.
It looks possible that it will be accepted and therefore we can’t expect
the attribute to be removed from govuk-frontend any time soon. Similarly
due to the pace WCAG moves we can’t expect the proposed changes to be
accepted any time soon. Therefore are forced to prevent this rule from
failing the tests until either govuk-frontend remove ‘aria-expanded’ from
radio buttons or WCAG accept that it should be included on radio buttons.