Ensure ‘make a decision’ radio buttons have unique IDs
This is a relatively major refactor.
Previously the index for the ids of each radio button was resetting with
each new `formPageRadio` macro meaning we had multiple buttons with the
id ‘decision’, ‘decision-1’ etc.
We can no longer use the macro but the HTML isn’t very verbose and
it solves a small usability issue that existed before