Stub entire application journey with Wiremock
Now we’re not saving the application in the session, we need a
way to stub the API in a stateful way. This uses Wiremock scenarios
(https://wiremock.org/docs/stateful-behaviour/) and the application’s
data to cycle through each task and page, and build up a scenario
and scenario states to return different data and set new states
when each scenario gets called.

This ensures a more realistic test, as we’re calling to our “real” API
each time, instead of relying on sessions, which can expire.