Add frontend JS and update template
This updates the template to add some frontend JS and simplify the
approach. We initially load 5 fieldsets, then hide all but one. If
the user needs more than 5 partners, we append to container. When we
show a new partner fieldset, we include a “remove” button, which
hides the containing fieldset, and clears out the input data.

There is an existing MOJ component (https://design-patterns.service.justice.gov.uk/components/add-another/),
and I’ve used some of the classes, but I felt like the needs that we
had were slightly different, and for accessiblity purposes, we needed
to have a number of fieldsets already visible.