Add placementDate response to the application data
Now we’re not saving the application to the session, a page doesn’t have
access to the updated application data. As `noticeTypeFromApplication`
relies on data from the current page AND data from the application, the
easiest way to get around this is to append the page body to the
application and then call `noticeTypeFromApplication` with the complete
response. I’ve added a test to make sure this happens and also documents
what we’re doing here.